backend-python/requirements.txt
# FastAPI Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# CORS
python-cors==1.0.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# Email (SendGrid)
sendgrid==6.11.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
